requirements.txt
beautifulsoup4==4.5.1
bleach==1.5.0
chardet==2.3.0
Django==1.10.7
django-contrib-comments==1.7.3
django-extensions==1.7.5
django-geoposition==0.3.0
django-taggit==0.21.3

future==0.16.0
grappelli-safe==0.4.5
html5lib==0.9999999
Mezzanine==4.2.3
oauthlib==2.0.1
Pillow==3.4.2
pytz==2016.10
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
tzlocal==1.3
whitenoise==3.2.3
django-watson==1.3.0
shortuuid==0.4.3
markdown==2.6.8
django-admin-sortable==2.0.21
geopy==1.11.0
easy-thumbnails==2.3
sorl-thumbnail==12.3
django_compressor==2.1.1
twython==3.4.0
django-embed-video==1.1.2
django-registration==2.2
django-storages-redux==1.3.2
py-bcrypt==0.4
mailchimp==2.0.9
newrelic==2.82.0.62

## Heroku
psycopg2
dj_database_url==0.4.2
redis==2.10.5
gunicorn==19.6
boto == 2.46.1
